WebUI/requirements.txt
flask>=3.0
flask-cors>=4.0
bcrypt>=4.1
PyJWT>=2.8
httpx>=0.27
gunicorn>=22.0
tencentcloud-sdk-python>=3.0.0
